Moved testing to main file
